% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{check_certara_package_versions}
\alias{check_certara_package_versions}
\title{Check if Certara packages are installed}
\usage{
check_certara_package_versions(pkgs)
}
\value{
Named character vector indicating package version, \code{NA} is returned if Certara package is not installed
}
\description{
Check if Certara packages are installed
}
\examples{
check_certara_package_versions()
}
