% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{check_installed_certara_packages}
\alias{check_installed_certara_packages}
\title{Check if Certara packages are installed}
\usage{
check_installed_certara_packages(pkgs)
}
\value{
Named logical vector indicating whether Certara package is installed
}
\description{
Check if Certara packages are installed
}
\examples{
check_installed_certara_packages()
}
