% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_certara_packages}
\alias{install_certara_packages}
\title{Install Certara R Packages}
\usage{
install_certara_packages()
}
\value{
\code{TRUE} if all packages successfully install
}
\description{
A user friendly wrapper to easily install R packages developed Certara and hosted on Certara's CRAN repo
}
