% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{remove_certara_packages}
\alias{remove_certara_packages}
\title{Remove Certara packages}
\usage{
remove_certara_packages()
}
\value{
\code{TRUE} if all packages successfully removed
}
\description{
A user friendly wrapper to easily remove R packages developed by Certara
}
\examples{
\dontrun{
remove_certara_packages()
}
}
