% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{certara_packages}
\alias{certara_packages}
\title{List all packages developed by Certara}
\usage{
certara_packages(include_self = FALSE, include_JFROG_deps = FALSE)
}
\arguments{
\item{include_self}{Logical; Set to \code{TRUE} to include \code{Certara.R}}

\item{include_JFROG_deps}{Logical; Set to \code{TRUE} to include Certara package dependencies on Jfrog}
}
\value{
Character vector of package names
}
\description{
List all packages developed by Certara
}
\examples{
certara_packages()
}
