% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{check_installed_certara_packages}
\alias{check_installed_certara_packages}
\title{Check if Certara packages are installed}
\usage{
check_installed_certara_packages(pkgs)
}
\arguments{
\item{pkgs}{Character or character vector of package names. If missing, defaults to \code{certara_packages()}}
}
\value{
Named logical vector indicating whether Certara package is installed
}
\description{
Check if Certara packages are installed
}
\examples{
\dontrun{
check_installed_certara_packages()
}
}
