% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_memory.R
\name{clear_memory}
\alias{clear_memory}
\title{Clear all memory records}
\usage{
clear_memory()
}
\value{
Invisibly \code{TRUE}.
}
\description{
Clear all memory records
}
\examples{
\dontrun{
clear_memory()
}
}
\keyword{internal}
