% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_certara_packages}
\alias{install_certara_packages}
\title{Install Certara R Packages}
\usage{
install_certara_packages(
  CRAN_repo = "https://cloud.r-project.org",
  clean_install = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{CRAN_repo}{Character;  The URL of a CRAN mirror which defaults to \code{"https://cloud.r-project.org"}}

\item{clean_install}{Logical; Set to \code{TRUE} to remove all existing Certara packages prior to installing}

\item{quiet}{Logical; Set to \code{TRUE} to suppress installation messages and reduce console output}
}
\value{
\code{TRUE} if all packages successfully install
}
\description{
A user friendly wrapper to easily install R packages and Shiny applications developed by Certara
}
